# docs/swagger.yaml
basePath: /
definitions:
  auth.DummyLoginRequest:
    properties:
      role:
        type: string
    type: object
  auth.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
  auth.RegisterRequest:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
  auth.TokenResponse:
    properties:
      token:
        type: string
    type: object
  booking.CreateBookingRequest:
    properties:
      slot_id:
        type: string
    type: object
  room.CreateRoomRequest:
    properties:
      capacity:
        type: integer
      description:
        type: string
      name:
        type: string
    type: object
  schedule.CreateScheduleRequest:
    properties:
      day_of_week:
        items:
          type: integer
        type: array
      end_time:
        type: string
      start_time:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: Сервис бронирования переговорок
  title: Room Booking Service
  version: "1.0"
paths:
  /bookings/{bookingId}/cancel:
    post:
      parameters:
      - description: ID брони
        in: path
        name: bookingId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Отменить бронь
      tags:
      - Bookings
  /bookings/create:
    post:
      consumes:
      - application/json
      parameters:
      - description: Данные брони
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/booking.CreateBookingRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Создать бронь
      tags:
      - Bookings
  /bookings/list:
    get:
      parameters:
      - description: Страница
        in: query
        name: page
        type: integer
      - description: Размер страницы
        in: query
        name: pageSize
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Все брони (admin)
      tags:
      - Bookings
  /bookings/my:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Мои брони
      tags:
      - Bookings
  /dummyLogin:
    post:
      consumes:
      - application/json
      parameters:
      - description: Роль
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/auth.DummyLoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/auth.TokenResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      summary: Получить тестовый JWT
      tags:
      - Auth
  /login:
    post:
      consumes:
      - application/json
      parameters:
      - description: Email и пароль
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/auth.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/auth.TokenResponse'
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      summary: Авторизация
      tags:
      - Auth
  /register:
    post:
      consumes:
      - application/json
      parameters:
      - description: Email и пароль
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/auth.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/auth.TokenResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      summary: Регистрация
      tags:
      - Auth
  /rooms/{roomId}/schedule/create:
    post:
      consumes:
      - application/json
      parameters:
      - description: ID комнаты
        in: path
        name: roomId
        required: true
        type: string
      - description: Данные расписания
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/schedule.CreateScheduleRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Создать расписание для комнаты
      tags:
      - Schedule
  /rooms/{roomId}/slots/list:
    get:
      parameters:
      - description: ID комнаты
        in: path
        name: roomId
        required: true
        type: string
      - description: Дата (YYYY-MM-DD)
        in: query
        name: date
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Получить слоты комнаты на дату
      tags:
      - Schedule
  /rooms/create:
    post:
      consumes:
      - application/json
      parameters:
      - description: Данные комнаты
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/room.CreateRoomRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Создать комнату
      tags:
      - Rooms
  /rooms/list:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Список всех комнат
      tags:
      - Rooms
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"